ark-log-watchdog/requirements.txt
mss
opencv-python
pillow
pytesseract
numpy
python-dotenv
PyYAML
customtkinter
requests 
pyjwt 
cryptography